Conditionally adjust priv-app path for lollipop+ devices
On Lollipop, the path to the Settings app is /system/priv-app/Settings/Settings.apk
This commit copies the data APK to a similarly-named path on devices with their ro.build.version.release's X.Y[.Z] version number's X
     greater than or equal to 5 (lollipop and up).
This is build property MAY not exist or have values of the nature assumed by this commit for all OEMs, but this definitely covers
     AOSP and popular open source derivatives there-of.

The backup script's list-files function now has BOTH possible priv-app paths listed.
This works because backup_file and restore_file no-op in the case that a file does not exist.