Change custom battery stats permission to normal & bump version to 1.6
This allows us to work better with the new Marshmallow permission model
by not requiring an explicit permission request.  The explicit request
works well when this is a known permission, but their model for custom
permissions is weak.  The fact that the user has installed the root
companion is good enough.